+++ HISTORY.rst
@@ -3,6 +3,10 @@ Changelog
 
 Unreleased
 ~~~~~~~~~~
+
+- Adds a status (unknown, interim, final) to elections and votes.
+  [msom]
+
 1.3.2 (2017-04-27)
 ~~~~~~~~~~~~~~~~~~~